requirements.txt
# Loqo AI Agent v3 — Dependencies
# Core
google-genai>=1.0.0
pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
langgraph>=0.2.0
python-dotenv>=1.0.0

# Server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Optional (uncomment for local evaluation models)
# transformers>=4.38.0
# torch>=2.1.0
# ultralytics>=8.1.0
